// src/app/gojs-diagram/enum/figure-type.enum.ts
export enum EnumFigureType {
	ND = 'None',
	RETTANGOLO = 'Rectangle',
	ARROTONDATO = 'RoundedRectangle',
	CERCHIO = 'Circle',
	FRECCIA_DESTRA = 'TriangleRight',
	FRECCIA_SINISTRA = 'TriangleLeft',
	FRECCIA_GIU = 'TriangleDown',
	FRECCIA_SU = 'TriangleUp',
	ROMBO = 'Diamond',
	ASTERISCO = 'AsteriskLine',
	PENTAGONO = 'Pentagon',
	ESAGONO = 'Esagon',
	STELLA_5 = 'FivePointedStar',
	STELLA_6 = 'SixPointedStar',
	SPLASH = 'NinePointedBurst',
	QUADRO = 'FramedRectangle',
	ANELLO = 'Ring',
	NUVOLA = 'Cloud',
	TORTA = 'Pie',
	PEZZO_TORZA = 'PiePiece',
	CROCE = 'ThinCross',
	CROCE_GRANDE = 'ThickCross',
	X = 'ThinX',
	X_GRANDE = 'ThickX',
	FILE = 'File',
}

// https://gojs.net/latest/samples/shapes.html